requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
biopython==1.85
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
et-xmlfile==2.0.0
executing==2.2.0
fonttools==4.58.0
frozenlist==1.6.0
httplib2==0.22.0
idna==3.10
ipython-pygments-lexers==1.1.1
jedi==0.19.2
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.0
